Major changes to Board.  Moving the getWord method from BoardTreePosition into this class, and adding some basic accessors (like getLetter).  Adding several methods to turn arrays of letters/ambiguity codes into WordPlay objects (these will help with Game loading, shortly).  Also adding an 'ambiguity' field, private, to help with figuring out which letter (when there are multiple options for a single letter) is being used based on the ambiguity codes, without having to do a linear search each time.
Mostly, though, the big change in this commit is the addition of the
enumerateWordPlays method, which takes a given Word and returns all the
ways (via a Collection of WordPlay objects) that that word could be
played on this board.

This is the last major missing requirement for a tree-search routine.